Create version 2 API of findMenusWithPaginationV2

+++ src/main/java/blowest/kiosk/dto/MenuPagedResponseDto.java
@@ -14,13 +14,13 @@ import java.util.List;
 @Getter
 public class MenuPagedResponseDto {
 
-    private int totalPages;
+    private long totalPages;
 
     private long totalMenus;
 
     List<MenuResponseDto> menus;
 
-    public static MenuPagedResponseDto create(int totalPages, long totalMenus, List<MenuResponseDto> menus) {
+    public static MenuPagedResponseDto create(long totalPages, long totalMenus, List<MenuResponseDto> menus) {
         MenuPagedResponseDto dto = new MenuPagedResponseDto();
         dto.totalPages = totalPages;
         dto.totalMenus = totalMenus;
